api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gentle-api"
version = "0.1.0"
description = "Gentle API - Backend for empathetic productivity companion"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "sqlalchemy>=2.0.0,<3.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary]>=3.1.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "python-dotenv>=1.0.0",
    "sentry-sdk>=1.38.0",
    "posthog>=3.0.0",
    "python-jose[cryptography]>=3.3.0",
    "cachetools>=5.3.0",
    "alembic>=1.12.0",
    "openai>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
select = ["E", "F", "I"]
line-length = 88
target-version = "py311"